+++ CHANGES.rst
@@ -1,10 +1,17 @@
+Version 0.2.2
+-------------
+
+Released 2020-09-29
+
+-   Add swagger index html
+-   Fix url of swagger JSON and url_prefix of swagger blueprint
+
 Version 0.2.1
 -------------
 
 Released 2020-09-28
 
 -   Fix remove-catch_all_404s from call of super class
-
 
 Version 0.2.0
 -------------